+++ transports/uds/src/lib.rs
@@ -114,7 +114,7 @@ impl Transport for $uds_config {
         }
     }
 
-    fn dial_with_role_override(self, addr: Multiaddr) -> Result<Self::Dial, TransportError<Self::Error>> {
+    fn dial_as_listener(self, addr: Multiaddr) -> Result<Self::Dial, TransportError<Self::Error>> {
         self.dial(addr)
     }